Remove old program detail page and Waffle switch gating the new page
The new page is released and stable. The switch and old code are no longer necessary.

ECOM-7482
