--- lms/static/js/learner_dashboard/program_details_factory_2017.js
@@ -1,13 +1,0 @@
-(function(define) {
-    'use strict';
-
-    define([
-        'js/learner_dashboard/views/program_details_view_2017'
-    ],
-    function(ProgramDetailsView) {
-        return function(options) {
-            var ProgramDetails = new ProgramDetailsView(options);
-            return ProgramDetails;
-        };
-    });
-}).call(this, define || RequireJS.define);
