+++ lms/static/js/learner_dashboard/views/program_header_view.js
@@ -2,17 +2,20 @@
     'use strict';
 
     define(['backbone',
-            'jquery',
-            'edx-ui-toolkit/js/utils/html-utils',
-            'text!../../../templates/learner_dashboard/program_header_view.underscore',
-            'picturefill'
-           ],
-         function(Backbone, $, HtmlUtils, pageTpl, picturefill) {
+        'jquery',
+        'edx-ui-toolkit/js/utils/html-utils',
+        'text!../../../templates/learner_dashboard/program_header_view.underscore',
+        'text!../../../images/programs/micromasters-program-details.svg',
+        'text!../../../images/programs/xseries-program-details.svg',
+        'text!../../../images/programs/professional-certificate-program-details.svg'
+    ],
+         function(Backbone, $, HtmlUtils, pageTpl, MicroMastersLogo,
+                  XSeriesLogo, ProfessionalCertificateLogo) {
              return Backbone.View.extend({
                  breakpoints: {
                      min: {
-                         'medium': '768px',
-                         'large': '1180px'
+                         medium: '768px',
+                         large: '1180px'
                      }
                  },
 
@@ -24,32 +27,29 @@
                      this.render();
                  },
 
+                 getLogo: function() {
+                     var logo = false,
+                         type = this.model.get('programData').type;
+
+                     if (type === 'MicroMasters') {
+                         logo = MicroMastersLogo;
+                     } else if (type === 'XSeries') {
+                         logo = XSeriesLogo;
+                     } else if (type === 'Professional Certificate') {
+                         logo = ProfessionalCertificateLogo;
+                     }
+                     return logo;
+                 },
+
                  render: function() {
                      var data = $.extend(this.model.toJSON(), {
-                         breakpoints: this.breakpoints
+                         breakpoints: this.breakpoints,
+                         logo: this.getLogo()
                      });
 
                      if (this.model.get('programData')) {
                          HtmlUtils.setHtml(this.$el, this.tpl(data));
-                         this.postRender();
                      }
-                 },
-
-                 postRender: function() {
-                    // To resolve a bug in IE with picturefill reevaluate images
-                     if (navigator.userAgent.indexOf('MSIE') !== -1 ||
-                        navigator.appVersion.indexOf('Trident/') > 0) {
-                        /* Microsoft Internet Explorer detected in. */
-                         window.setTimeout(function() {
-                             this.reEvaluatePicture();
-                         }.bind(this), 100);
-                     }
-                 },
-
-                 reEvaluatePicture: function() {
-                     picturefill({
-                         reevaluate: true
-                     });
                  }
              });
          }
