+++ lms/static/js/learner_dashboard/views/upgrade_message_view.js
@@ -1,25 +1,22 @@
 (function(define) {
     'use strict';
     define(['backbone',
-            'jquery',
-            'underscore',
-            'gettext',
-            'edx-ui-toolkit/js/utils/html-utils',
-            'text!../../../templates/learner_dashboard/upgrade_message.underscore',
-            'text!../../../templates/learner_dashboard/certificate_icon.underscore'
-           ],
+        'jquery',
+        'underscore',
+        'gettext',
+        'edx-ui-toolkit/js/utils/html-utils',
+        'text!../../../templates/learner_dashboard/upgrade_message.underscore'
+    ],
          function(
              Backbone,
              $,
              _,
              gettext,
              HtmlUtils,
-             upgradeMessageTpl,
-             certificateIconTpl
+             upgradeMessageTpl
          ) {
              return Backbone.View.extend({
                  messageTpl: HtmlUtils.template(upgradeMessageTpl),
-                 iconTpl: HtmlUtils.template(certificateIconTpl),
 
                  initialize: function(options) {
                      this.$el = options.$el;
@@ -29,7 +26,6 @@
                  render: function() {
                      var data = this.model.toJSON();
 
-                     data = $.extend(data, {certificateSvg: this.iconTpl()});
                      HtmlUtils.setHtml(this.$el, this.messageTpl(data));
                  }
              });
