+++ lms/templates/learner_dashboard/certificate_status.underscore
@@ -1,12 +1,5 @@
-<div class="message col-12 md-col-8">
-    <% // safe-lint: disable=underscore-not-escaped %>
-    <span class="certificate-icon green-icon" aria-hidden="true"><%= certificateSvg %></span>
-    <span class="card-msg"><%- gettext('Congratulations! You have earned a certificate for this course.') %></span>
-</div>
-<div class="action col-12 md-col-4"> 
-    <a href="<%- certificate_url %>" class="btn-brand cta-secondary">
-        <% // safe-lint: disable=underscore-not-escaped %>
-        <span class="certificate-icon blue-icon" aria-hidden="true"><%= certificateSvg %></span>
-        <%- gettext('View Certificate') %>
-    </a>
-</div>
+<p class="certificate-status col-12 md-col-8">
+    <span class="card-msg"><%- gettext('Certificate Status:') %></span>
+    <span class="fa fa-check-circle" aria-hidden="true"></span>
+    <span class="certificate-status-msg"><%- gettext('Certificate Purchased') %></span>
+</p>
