--- lms/templates/learner_dashboard/certificate_status_2017.underscore
@@ -1,5 +1,0 @@
-<p class="certificate-status col-12 md-col-8">
-    <span class="card-msg"><%- gettext('Certificate Status:') %></span>
-    <span class="fa fa-check-circle" aria-hidden="true"></span>
-    <span class="certificate-status-msg"><%- gettext('Certificate Purchased') %></span>
-</p>
