+++ lms/templates/learner_dashboard/course_card.underscore
@@ -1,21 +1,7 @@
 <div class="section">
     <div class="course-meta-container col-12 md-col-8 sm-col-12">
-        <div class="course-image-container">
-            <% if ( marketing_url || course_url ) { %>
-                <a href="<%- marketing_url || course_url %>" class="course-image-link">
-                    <img
-                        class="header-img"
-                        src="<%- course_image_url %>"
-                        <% // safe-lint: disable=underscore-not-escaped %>
-                        alt="<%= interpolate(gettext('%(courseName)s Home Page.'), {courseName: title}, true) %>"/>
-                </a>
-            <% } else { %>
-                <img class="header-img" src="<%- course_image_url %>" alt=""/>
-            <% } %>
-
-        </div>
         <div class="course-details">
-            <h3 class="course-title">
+            <h5 class="course-title">
                 <% if ( marketing_url || course_url ) { %>
                     <a href="<%- marketing_url || course_url %>" class="course-title-link">
                         <%- title %>
@@ -23,19 +9,20 @@
                 <% } else { %>
                     <%- title %>
                 <% } %>
-            </h3>
+            </h5>
             <div class="course-text">
-                <% if (start_date && end_date) { %>
-                    <span class="run-period"><%- start_date %> - <%- end_date %></span>
-                    -
+                <% if (enrolled) { %>
+                    <span class="enrolled"><%- enrolled %>: </span>
                 <% } %>
-                <span class="course-key"><%- course_key %></span>
+                <% if (dateString) { %>
+                    <span class="run-period"><%- dateString %></span>
+                <% } %>
             </div>
         </div>
-    </div>
-    <div class="course-actions col-12 md-col-4 sm-col-12">        
+        <div class="course-actions col-12 md-col-4 sm-col-12"></div>
+        <div class="course-certificate certificate-status"></div>
     </div>
 </div>
 <div class="section action-msg-view"></div>
 <div class="section upgrade-message"></div>
-<div class="section certificate-status"></div>
+<div class="section expired-notification"></div>
