--- lms/templates/learner_dashboard/course_card_2017.underscore
@@ -1,28 +1,0 @@
-<div class="section">
-    <div class="course-meta-container col-12 md-col-8 sm-col-12">
-        <div class="course-details">
-            <h5 class="course-title">
-                <% if ( marketing_url || course_url ) { %>
-                    <a href="<%- marketing_url || course_url %>" class="course-title-link">
-                        <%- title %>
-                    </a>
-                <% } else { %>
-                    <%- title %>
-                <% } %>
-            </h5>
-            <div class="course-text">
-                <% if (enrolled) { %>
-                    <span class='enrolled'><%- enrolled %>: </span>
-                <% } %>
-                <% if (dateString) { %>
-                    <span class="run-period"><%- dateString %></span>
-                <% } %>
-            </div>
-        </div>
-        <div class="course-actions col-12 md-col-4 sm-col-12"></div>
-        <div class="course-certificate certificate-status"></div>
-    </div>
-</div>
-<div class="section action-msg-view"></div>
-<div class="section upgrade-message"></div>
-<div class="section expired-notification"></div>
