+++ lms/templates/learner_dashboard/course_enroll.underscore
@@ -1,29 +1,19 @@
-<% if (is_enrolled) { %>
-    <div class="enrollment-info"><%- gettext('enrolled') %></div>
-    <% if (is_enrollment_open || is_course_ended) { %>
-        <a href="<%- course_url %>" class="btn view-course-link">
-            <% if (is_enrollment_open) { %>
-                <%- gettext('View Course') %>
-            <% } else if (is_course_ended) { %>
-                <%- gettext('View Archived Course') %>
-            <% } %>
-        </a>
-    <% } %>
+<% if (is_enrolled && (typeof expired === 'undefined' || expired === false)) { %>
+    <a href="<%- course_url %>" class="view-course-button btn-brand btn cta-primary">
+        <% if (is_course_ended) { %>
+            <%- gettext('View Archived Course') %>
+        <% } else { %>
+            <%- gettext('View Course') %>
+        <% } %>
+    </a>
 <% } else { %>
     <% if (enrollable_course_runs.length > 0) { %>
-        <div class="enrollment-info"><%- gettext('Not Enrolled') %></div>
         <% if (enrollable_course_runs.length > 1) { %>
             <div class="run-select-container">
-                <div class="select-error">
-                    <%- gettext('Please select a course date') %>
-                </div>
-                <label class="sr-only" for="select-<%- course_key %>-run">
-                    <%- gettext('Select Course Run') %>
+                <label class="select-choice" for="select-<%- course_key %>-run">
+                    <%- gettext('Choose a course run:') %>
                 </label>
-                <select id="select-<%- course_key %>-run" class="run-select" autocomplete="off">
-                    <option value="" selected="selected">
-                        <%- gettext('Choose Course Date') %>
-                    </option>
+                <select id="select-<%- course_key %>-run" class="run-select field-input input-select" autocomplete="off">
                     <% _.each (enrollable_course_runs, function(courseRun) { %>
                         <option
                             value="<%- courseRun.key %>"
@@ -31,19 +21,17 @@
                                 selected="selected"
                             <% }%>
                         >
-                            <%= interpolate(
-                                gettext('Starts %(start)s'),
-                                { start: courseRun.start_date },
-                                true)
-                            %>
+                            <%- courseRun.dateString %>
                         </option>
                     <% }); %>
                 </select>
             </div>
         <% } %>
-        <button type="button" class="btn-brand btn cta-primary enroll-button">
-            <%- gettext('Enroll Now') %>
-        </button>
+        <div class="enroll-button">
+            <button type="button" class="btn-brand btn cta-primary">
+                <%- gettext('Enroll Now') %>
+            </button>
+        </div>
     <% } else if (upcoming_course_runs.length > 0) {%>
         <div class="no-action-message">
             <%- gettext('Coming Soon') %>
