+++ lms/templates/learner_dashboard/program_details.html
@@ -15,6 +15,8 @@ from openedx.core.djangolib.js_utils import (
 <%static:require_module module_name="js/learner_dashboard/program_details_factory" class_name="ProgramDetailsFactory">
 ProgramDetailsFactory({
     programData: ${program_data | n, dump_js_escaped_json},
+    courseData: ${course_data | n, dump_js_escaped_json},
+    certificateData: ${certificate_data | n, dump_js_escaped_json},
     urls: ${urls | n, dump_js_escaped_json},
     userPreferences: ${user_preferences | n, dump_js_escaped_json},
 });
@@ -24,8 +26,6 @@ ProgramDetailsFactory({
 <%block name="pagetitle">${_("Program Details")}</%block>
 <%block name="bodyclass">program-details</%block>
 
-<%include file="_dashboard_navigation_programs.html"/>
-
 <main id="main" aria-label="Content" tabindex="-1">
-    <div class="js-program-details-wrapper"></div>
+    <div class="js-program-details-wrapper program-details-wrapper"></div>
 </main>
