--- lms/templates/learner_dashboard/program_details_2017.html
@@ -1,31 +1,0 @@
-## Override the default styles_version to the Pattern Library version (version 2)
-<%! main_css = "style-learner-dashboard" %>
-
-<%page expression_filter="h"/>
-<%inherit file="../main.html" />
-<%namespace name='static' file='../static_content.html'/>
-<%!
-from django.utils.translation import ugettext as _
-from openedx.core.djangolib.js_utils import (
-    dump_js_escaped_json, js_escaped_string
-)
-%>
-
-<%block name="js_extra">
-<%static:require_module module_name="js/learner_dashboard/program_details_factory_2017" class_name="ProgramDetailsFactory2017">
-ProgramDetailsFactory2017({
-    programData: ${program_data | n, dump_js_escaped_json},
-    courseData: ${course_data | n, dump_js_escaped_json},
-    certificateData: ${certificate_data | n, dump_js_escaped_json},
-    urls: ${urls | n, dump_js_escaped_json},
-    userPreferences: ${user_preferences | n, dump_js_escaped_json},
-});
-</%static:require_module>
-</%block>
-
-<%block name="pagetitle">${_("Program Details")}</%block>
-<%block name="bodyclass">program-details</%block>
-
-<main id="main" aria-label="Content" tabindex="-1">
-    <div class="js-program-details-wrapper program-details-wrapper"></div>
-</main>
