--- lms/templates/learner_dashboard/upgrade_message_2017.underscore
@@ -1,10 +1,0 @@
-<div class="message certificate-status col-12 md-col-8">
-    <span class="card-msg"><%- gettext('Certificate Status:') %></span>
-    <span><%- gettext('Needs verified certificate ') %></span>
-    <span class="price">  <%- price %></span>
-</div>
-<div class="action col-12 md-col-4">
-    <a href="<%- upgrade_url %>" class="btn-brand btn cta-primary upgrade-button">
-        <%- gettext('Buy Certificate') %>
-    <a>
-</div>
